reorganize cmdline to avoid loading deps
