+++ openst/preprocessing/image_preprocess.py
@@ -1,4 +1,3 @@
-import argparse
 import numpy as np
 import os
 
@@ -6,47 +5,6 @@ from PIL import Image
 from tqdm import tqdm
 
 # from openst.preprocessing.CUT.models import create_model
-
-def get_image_preprocess_parser():
-    """
-    Parse command-line arguments.
-
-    Returns:
-        argparse.Namespace: Parsed command-line arguments.
-    """
-    parser = argparse.ArgumentParser(
-        description="preprocess imaging data with CUT model (as in Open-ST paper)",
-        allow_abbrev=False,
-        add_help=False,
-    )
-    parser.add_argument('--input_img', type=str, required=True,
-                        help='path to input image')
-    parser.add_argument('--cut_dir', type=str, required=True,
-                        help='path to CUT directory (to save patched images)')
-    parser.add_argument('--tile_size_px', type=int, required=True,
-                        help='size of the tile in pixels')
-    parser.add_argument(
-        "--device",
-        type=str,
-        default="cpu",
-        choices=["cpu", "cuda"],
-        help="Device used to run feature matching model. Can be ['cpu', 'cuda']",
-    )
-    parser.add_argument('--checkpoints_dir', type=str, default='./checkpoints', help='models are saved here')
-
-    return parser
-
-
-def setup_image_preprocess_parser(parent_parser):
-    """setup_image_preprocess_parser"""
-    parser = parent_parser.add_parser(
-        "image_preprocess",
-        help="convert fastq files into spatial barcode files (sequence and coordinates)",
-        parents=[get_image_preprocess_parser()],
-    )
-    parser.set_defaults(func=_run_image_preprocess)
-
-    return parser
 
 def _images_to_tiles(img):
     _img_shape = img.shape
@@ -108,5 +66,6 @@ def _run_image_preprocess():
     Image.fromarray(img_restitch.astype(np.uint8)).save(args.output_img)
 
 if __name__ == "__main__":
+    from openst.cli import get_image_preprocess_parser
     args = get_image_preprocess_parser().parse_args()
     _run_image_preprocess(args)
