+++ openst/segmentation/segment.py
@@ -1,4 +1,3 @@
-import argparse
 import h5py
 import logging
 import dask
@@ -26,193 +25,6 @@ from openst.utils.pseudoimage import create_unpaired_pseudoimage
 from skimage.segmentation import find_boundaries
 
 # TODO: implement gray_dilation
-
-
-def get_segment_parser():
-    """
-    Parse command-line arguments.
-
-    Returns:
-        argparse.Namespace: Parsed command-line arguments.
-    """
-    parser = argparse.ArgumentParser(
-        description="segmentation of Open-ST imaging data with cellpose",
-        allow_abbrev=False,
-        add_help=False,
-    )
-
-    # Data
-    parser.add_argument(
-        "--image-in",
-        type=str,
-        help="Path to the input image.",
-    )
-    parser.add_argument(
-        "--rna-segment",
-        action="store_true",
-        help="""Performs segmentation based on local RNA density pseudoimages from sequencing data,
-              instead of using a staining image. 
-              This assumes coordinates in microns (can be transformed with --rna-segment-input-resolution)"""
-    )
-    parser.add_argument(
-        "--output-mask",
-        type=str,
-        required=True,
-        help="Path to the output file where the mask will be stored",
-    )
-    parser.add_argument(
-        "--adata",
-        type=str,
-        default="",
-        help="When specified, staining image is loaded from adata (from --image-in), and segmentation is saved there (to --output-mask)",
-    )
-    
-    # Cellpose
-    parser.add_argument(
-        "--model",
-        type=str,
-        default="",
-        help="""cellpose model - either a path or a valid string to pretrained model.""",
-    )
-    parser.add_argument(
-        "--flow-threshold",
-        type=float,
-        default=0.5,
-        help="cellpose's 'flow_threshold' parameter",
-    )
-    parser.add_argument(
-        "--cellprob-threshold",
-        type=float,
-        default=0,
-        help="cellpose's 'cellprob_threshold' parameter",
-    )
-    parser.add_argument(
-        "--diameter",
-        type=float,
-        default=20,
-        help="cellpose's 'diameter' parameter",
-    )
-    parser.add_argument(
-        "--chunk-size",
-        type=int,
-        default=512,
-        help="When prediction of the mask runs in separate chunks, this is the chunk square size (in pixels)",
-    )
-    parser.add_argument(
-        "--chunked",
-        action="store_true",
-        help="When specified, segmentation is computed at non-overlapping chunks of size '--chunk-size'",
-    )
-    parser.add_argument(
-        "--max-image-pixels",
-        type=int,
-        default=933120000,
-        help="Upper bound for number of pixels in the images (prevents exception when opening very large images)",
-    )
-    parser.add_argument(
-        "--gpu",
-        action="store_true",
-        help="When specified, a GPU will be used for prediction",
-    )
-
-    # Postprocessing
-    parser.add_argument(
-        "--dilate-px",
-        type=int,
-        help="Pixels the outlines of the segmentation mask will be extended",
-        required=False,
-        default=0,
-    )
-    parser.add_argument(
-        "--outline-px",
-        type=int,
-        help="Objects will be represented as px-width outlines (only if >0)",
-        required=False,
-        default=0,
-    )
-
-    # Preprocessing
-    parser.add_argument(
-        "--mask-tissue",
-        action="store_true",
-        help="Tissue (imaging modality) is masked from the background before segmentation",
-    )
-    parser.add_argument(
-        "--tissue-masking-gaussian-sigma",
-        type=int,
-        default=5,
-        help="The gaussian blur sigma used during the isolation of the tissue on the staining image",
-    )
-    parser.add_argument(
-        "--keep-black-background",
-        action="store_true",
-        help="Whether to set the background of the imaging modalities to white after tissue masking",
-    )
-
-    # RNA density-based segmentation configuration
-    parser.add_argument(
-        "--rna-segment-spatial-coord-key",
-        type=str,
-        default='obsm/spatial',
-        help="Path to the spatial coordinates inside the AnnData object (e.g., 'obsm/spatial')"
-    )
-    parser.add_argument(
-        "--rna-segment-input-resolution",
-        type=float,
-        default=1,
-        help="""Spatial resolution of the input coordinates (retrieved from --rna-segment-spatial-coord-key).
-              If it is in microns, leave as 1. If it is in pixels, specify the pixel to micron conversion factor."""
-    )
-    parser.add_argument(
-        "--rna-segment-render-scale",
-        type=float,
-        default=2,
-        help="Size of bins for computing the binning (in microns). For Open-ST v1, we recommend a value of 2."
-    )
-    parser.add_argument(
-        "--rna-segment-render-sigma",
-        type=float,
-        default=1,
-        help="Smoothing factor applied to the RNA pseudoimage (higher values lead to smoother images)"
-    )
-    parser.add_argument(
-        "--rna-segment-output-resolution",
-        type=float,
-        default=0.6,
-        help="Final resolution (micron/pixel) for the segmentation mask."
-    )
-    
-
-    # General
-    parser.add_argument(
-        "--num-workers",
-        type=int,
-        help="Number of parallel workers when --chunked is specified",
-        required=False,
-        default=-1,
-    )
-    parser.add_argument(
-        "--metadata-out",
-        type=str,
-        default="",
-        help="""Path where the metadata will be stored.
-        If not specified, metadata is not saved.
-        Warning: a report (via openst report) cannot be generated without metadata!""",
-    )
-    return parser
-
-
-def setup_segment_parser(parent_parser):
-    """setup_segment_parser"""
-    parser = parent_parser.add_parser(
-        "segment",
-        help="segmentation of Open-ST imaging data with cellpose",
-        parents=[get_segment_parser()],
-    )
-    parser.set_defaults(func=_run_segment)
-
-    return parser
-
 
 def assemble_from_tiles(tiles, width: int, height: int, channels: int, tile_size: int = 512):
     """
@@ -542,5 +354,6 @@ def _run_segment(args):
     _save_mask(adata, im, mask_complete, args)
 
 if __name__ == "__main__":
+    from openst.cli import get_segment_parser
     args = get_segment_parser().parse_args()
     _run_segment()
